DRILL-6888: Move nested classes outside HashAggTemplate to allow for plain java compile option closes #1569
